chore(ui): move loader to ui folder
