+++ packages/nextjs/ui/index.ts
@@ -4,5 +4,6 @@ export { Header } from "./Header";
 export { Footer } from "./Footer";
 export { Input } from "./Input/Input";
 export { LargeLogo } from "./LargeLogo";
+export { Loader } from "./Loader";
 export { Text } from "./Text/Text";
 export { Row } from "./Row";
